create pack str utility
