+++ src/a08_bud_atom_logic/atom.py
@@ -22,20 +22,22 @@ from src.a03_group_logic.group import awardlink_shop
 from src.a05_item_logic.item import itemunit_shop
 from src.a06_bud_logic.bud import BudUnit
 from src.a06_bud_logic.bud_tool import bud_attr_exists, bud_get_obj
-from src.a08_bud_atom_logic.atom_config import (
-    get_dimen_from_dict,
-    get_atom_config_jkeys,
+from src.a08_bud_atom_logic._utils.str_a08 import (
     atom_delete,
     atom_insert,
     atom_update,
     atom_hx_table_name,
+    CRUD_command,
+)
+from src.a08_bud_atom_logic.atom_config import (
+    get_dimen_from_dict,
+    get_atom_config_jkeys,
     get_atom_order,
     get_atom_config_dict,
     is_bud_dimen,
     get_atom_config_args,
     get_sorted_jkey_keys,
     get_atom_args_class_types,
-    CRUD_command,
 )
 from dataclasses import dataclass
