+++ src/a08_bud_atom_logic/test_/test_atom_core.py
@@ -10,7 +10,7 @@ from src.a06_bud_logic._utils.str_a06 import (
     credit_belief_str,
     debtit_belief_str,
 )
-from src.a08_bud_atom_logic.atom_config import atom_insert, atom_delete
+from src.a08_bud_atom_logic._utils.str_a08 import atom_insert, atom_delete
 from src.a08_bud_atom_logic.atom import BudAtom, budatom_shop
