+++ src/a08_bud_atom_logic/test_/test_atom_json.py
@@ -1,7 +1,7 @@
 from src.a01_word_logic.road import create_road
 from src.a06_bud_logic._utils.str_a06 import bud_item_factunit_str
 from src.a06_bud_logic._utils.str_a06 import road_str, base_str
-from src.a08_bud_atom_logic.atom_config import (
+from src.a08_bud_atom_logic._utils.str_a08 import (
     jkeys_str,
     jvalues_str,
     dimen_str,
