+++ src/a09_pack_logic/_utils/example_atoms.py
@@ -1,7 +1,7 @@
 from src.a01_word_logic.road import create_road, FiscTag
 from src.a06_bud_logic._utils.str_a06 import bud_itemunit_str
 from src.a06_bud_logic._utils.str_a06 import parent_road_str, item_tag_str
-from src.a08_bud_atom_logic.atom_config import atom_insert
+from src.a08_bud_atom_logic._utils.str_a08 import atom_insert
 from src.a08_bud_atom_logic.atom import budatom_shop, BudAtom
