+++ src/a09_pack_logic/_utils/example_deltas.py
@@ -3,7 +3,7 @@ from src.a06_bud_logic._utils.str_a06 import (
     bud_acctunit_str,
     acct_name_str,
 )
-from src.a08_bud_atom_logic.atom_config import atom_delete, atom_update
+from src.a08_bud_atom_logic._utils.str_a08 import atom_delete, atom_update
 from src.a08_bud_atom_logic.atom import budatom_shop
 from src.a09_pack_logic.delta import BudDelta, buddelta_shop
