+++ src/a09_pack_logic/delta.py
@@ -17,7 +17,7 @@ from src.a03_group_logic.acct import MemberShip, AcctName, AcctUnit
 from src.a03_group_logic.group import MemberShip
 from src.a05_item_logic.item import ItemUnit
 from src.a06_bud_logic.bud import BudUnit, budunit_shop
-from src.a08_bud_atom_logic.atom_config import CRUD_command
+from src.a08_bud_atom_logic._utils.str_a08 import CRUD_command
 from src.a08_bud_atom_logic.atom import (
     BudAtom,
     budatom_shop,
