+++ src/a09_pack_logic/test_delta/test_delta_legible_itemunit.py
@@ -11,7 +11,7 @@ from src.a06_bud_logic._utils.str_a06 import (
     morph_str,
     mass_str,
 )
-from src.a08_bud_atom_logic.atom_config import atom_update, atom_insert, atom_delete
+from src.a08_bud_atom_logic._utils.str_a08 import atom_update, atom_insert, atom_delete
 from src.a08_bud_atom_logic.atom import budatom_shop
 from src.a09_pack_logic.delta import buddelta_shop
 from src.a09_pack_logic.legible import create_legible_list
