+++ src/a10_bud_calc/test_/test_bud_calc_config.py
@@ -48,13 +48,13 @@ from src.a06_bud_logic._utils.str_a06 import (
     type_TagUnit_str,
     type_RoadUnit_str,
 )
-from src.a08_bud_atom_logic.atom_config import (
-    get_atom_config_dict,
+from src.a08_bud_atom_logic._utils.str_a08 import (
     class_type_str,
     jkeys_str,
     jvalues_str,
     sqlite_datatype_str,
 )
+from src.a08_bud_atom_logic.atom_config import get_atom_config_dict
 from src.a10_bud_calc.bud_calc_config import (
     jmetrics_str,
     fund_take_str,
