+++ src/a13_bud_listen_logic/_utils/example_listen_atoms.py
@@ -10,7 +10,7 @@ from src.a06_bud_logic._utils.str_a06 import (
     fopen_str,
     fnigh_str,
 )
-from src.a08_bud_atom_logic.atom_config import atom_insert, atom_update, atom_delete
+from src.a08_bud_atom_logic._utils.str_a08 import atom_insert, atom_update, atom_delete
 from src.a08_bud_atom_logic.atom import budatom_shop, BudAtom
 from src.a09_pack_logic.delta import buddelta_shop, BudDelta
