+++ src/a15_fisc_logic/test_fisc_/test_journal_sqlstr_.py
@@ -1,7 +1,7 @@
 from src.a01_word_logic.road import create_road
 from src.a06_bud_logic._utils.str_a06 import bud_item_factunit_str
 from src.a06_bud_logic._utils.str_a06 import fopen_str, base_str, road_str
-from src.a08_bud_atom_logic.atom_config import atom_insert, atom_hx_table_name
+from src.a08_bud_atom_logic._utils.str_a08 import atom_insert, atom_hx_table_name
 from src.a08_bud_atom_logic.atom import budatom_shop
 from src.a15_fisc_logic.journal_sqlstr import (
     get_atom2delta_table_create_sqlstr,
