+++ src/a17_idea_logic/test_idea/test_idea__config.py
@@ -62,12 +62,7 @@ from src.a07_calendar_logic._utils.str_a07 import (
     timeline_tag_str,
     yr1_jan1_offset_str,
 )
-from src.a08_bud_atom_logic.atom_config import (
-    get_atom_args_dimen_mapping,
-    get_atom_config_dict,
-    get_bud_dimens,
-    get_delete_key_name,
-    get_all_bud_dimen_delete_keys,
+from src.a08_bud_atom_logic._utils.str_a08 import (
     jkeys_str,
     jvalues_str,
     normal_specs_str,
@@ -75,6 +70,13 @@ from src.a08_bud_atom_logic.atom_config import (
     atom_delete,
     atom_insert,
     atom_update,
+)
+from src.a08_bud_atom_logic.atom_config import (
+    get_atom_args_dimen_mapping,
+    get_atom_config_dict,
+    get_bud_dimens,
+    get_delete_key_name,
+    get_all_bud_dimen_delete_keys,
 )
 from src.a10_bud_calc.bud_calc_config import (
     get_all_bud_calc_args,
