+++ src/a17_idea_logic/test_idea/test_idea_format_generate_idea.py
@@ -16,7 +16,7 @@ from src.a06_bud_logic._utils.str_a06 import (
     credit_vote_str,
 )
 from src.a06_bud_logic._utils.example_buds import budunit_v001
-from src.a08_bud_atom_logic.atom_config import atom_insert
+from src.a08_bud_atom_logic._utils.str_a08 import atom_insert
 from src.a08_bud_atom_logic.atom import budatom_shop
 from src.a17_idea_logic.idea import create_idea_df, make_buddelta, get_idearef_obj
 from src.a17_idea_logic.idea_config import (
